move fixture loader to source module to allow usage in both source and test modules
